api/requirements/requirements.txt
# Core
pytz
Django==4.2
gunicorn
Pillow

# Database
psycopg2-binary==2.9.9
dj-database-url

# Model tools
django-model-utils
django_unique_upload

# REST API
djangorestframework
Markdown
django-filter

# Auth — JWT (djoser for user management, simplejwt for tokens)
djoser
djangorestframework-simplejwt

# OpenAPI schema
drf-spectacular

# Storage
django-storages==1.14.2
boto3
minio

# CORS
django-cors-headers

# Environment
environs
django-configurations

# Task queue (Redis-backed)
arq
redis

# Logging
structlog

# Rich text
django-richtextfield

# JSON-LD
pyld
